% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rapprox_unif}
\alias{rapprox_unif}
\title{Draw Random Numbers from Approximated (Differentiably) Uniform Distribution.}
\usage{
rapprox_unif(n = 100, scale, tildec = 13.86294, seed = 123)
}
\arguments{
\item{n}{number of draws.}

\item{scale}{the scale parameter originally defining the upper bound of the uniform distribution.}

\item{tildec}{denotes the ratio between scale parameter \eqn{\theta} and \eqn{s}. The latter is responsible for the closeness of
the approximation to the uniform distribution. See also below for further details and the default value.}

\item{seed}{denotes the seed}
}
\value{
n draws with density \code{\link{papprox_unif}}.
}
\description{
Draw Random Numbers from Approximated (Differentiably) Uniform Distribution.
}
\details{
The method is based on the inversion method and the quantile function is computed numerically using \code{\link{uniroot}}.
}
\author{
Nadja Klein
}
\references{
Nadja Klein and Thomas Kneib (2015). Scale-Dependent Priors for Variance Parameters in Structured Additive Distributional Regression.
\emph{Working Paper}.
}
\seealso{
\code{\link{rapprox_unif}},\code{\link{papprox_unif}}
}

