% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sdPrior.r
\docType{data}
\name{zambia_graph}
\alias{zambia_graph}
\title{Prior precision  matrix for spatial variable in Zambia data set}
\description{
This is a 57x57 matrix containing row- and columwise the regions of Zambia, and the entries
 define the neighbourhoodstructure. The corresponding map sambia.bnd can be downloaded from \url{http://www.stat.uni-muenchen.de/~kneib/regressionsbuch/daten_e.html}.
 from the bnd file the prior precision matrix is obtained by
 library(BayesX)
 map <- read.bnd("zambia.bnd")
 K <- bnd2gra(map)
}
