% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{zambia_height92}
\alias{zambia_height92}
\title{Malnutrition in Zambia}
\format{A data frame with 4421 rows and 21 variables}
\source{
\url{http://www.stat.uni-muenchen.de/~kneib/regressionsbuch/daten_e.html}
}
\description{
The primary goal of a statistical analysis is to determine the effect of certain socioeconomic
             variables of the child, the mother, and the household on the child`s nutritional
             condition
             \itemize{
               \item zscore child`s Z-score
               \item c _breastf duration of breastfeeding in months
               \item c_age child`s age in months
               \item m_agebirth mother`s age at birth in years
               \item m_height mother`s height in centimeter
               \item m_bmi mother`s body mass index
               \item m_education mother`s level of education
               \item m_work mother`s work status
               \item region region of residence in Zambia
               \item district district of residence in Zambia
             }
}

